refactor: ChartType定義をviewTypes.tsに移動
